+++ src/Querier.SqlQuery/Interfaces/IBaseQuery.cs
@@ -3,7 +3,16 @@
 namespace Querier.SqlQuery.Interfaces
 {
     public interface IBaseQuery<TQuery> :
-        IFromSqlQuery<TQuery>, ISelectSqlQuery<TQuery>, IJoinSqlQuery<TQuery>, IWhereSqlQuery<TQuery>, IWhereShorthandSqlQuery<TQuery>, IGroupSqlQuery<TQuery>, IUnionSqlQuery<TQuery>, IOrderSqlQuery<TQuery> where TQuery : IBaseQuery<TQuery>
+        IFromSqlQuery<TQuery>,
+        ICteSqlQuery<TQuery>,
+        ISelectSqlQuery<TQuery>,
+        IJoinSqlQuery<TQuery>,
+        IWhereSqlQuery<TQuery>,
+        IWhereShorthandSqlQuery<TQuery>,
+        IGroupSqlQuery<TQuery>,
+        IUnionSqlQuery<TQuery>,
+        IOrderSqlQuery<TQuery>,
+        IRawSqlQuery<TQuery> where TQuery : IBaseQuery<TQuery>
     {
         TQuery New();
         SqlQueryResult Compile();