+++ src/Querier.SqlQuery/Interfaces/ISelectSqlQuery.cs
@@ -21,7 +21,10 @@
         TQuery SelectMonth(string column, string? columnAs = null);
         TQuery SelectYear(string column, string? columnAs = null);
 
-        TQuery SelectRaw(string sql, string sqlAs);
+        TQuery SelectRaw(string sql, string? sqlAs = null);
+
+        TQuery SelectCoalesce<T>(Func<TQuery, TQuery> query, T value, string? queryAs = null);
+        TQuery SelectCoalesceRaw(Func<TQuery, TQuery> query, string rawSql, string? queryAs = null);
 
         TQuery Distinct();
     }