+++ tests/Querier.SqlQuery.MySql.Tests/Unit/SqlJoinTests.cs
@@ -1,9 +1,4 @@
 ﻿using Querier.SqlQuery.Functions;
-using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
 
 namespace Querier.SqlQuery.MySql.Tests.Unit
 {
@@ -19,11 +14,11 @@ namespace Querier.SqlQuery.MySql.Tests.Unit
         [Fact]
         public void OrderBy()
         {
-            const string join1 = "select * from `orders` as `orders` inner join `products` on `products`.`id` = `orders`.`product_id`";
-            const string join2 = "select * from `orders` as `orders` inner join `products` on `products`.`id` = `orders`.`product_id` inner join `categories` on `categories`.`id` = `orders`.`category_id`";
+            const string join1 = "select * from `orders` as `orders` inner join `products` on `orders`.`product_id` = `products`.`id`";
+            const string join2 = "select * from `orders` as `orders` inner join `products` on `orders`.`product_id` = `products`.`id` inner join `categories` on `orders`.`category_id` = `categories`.`id`";
 
-            Assert.Equal(join1, _query.New().From("orders").Join("products", "id", "product_id").Compile().CompiledSql);
-            Assert.Equal(join2, _query.New().From("orders").Join("products", "id", "product_id").Join("categories", "id", "category_id").Compile().CompiledSql);
+            Assert.Equal(join1, _query.New().From("orders").Select().Join("product_id", "products", "id").Compile().CompiledSql);
+            Assert.Equal(join2, _query.New().From("orders").Select().Join("product_id", "products", "id").Join("category_id", "categories", "id").Compile().CompiledSql);
         }
     }
 }